rename SceneManager -> App
